requirements.txt
Django
django-bootstrap4
django-allauth
django-ckeditor
